requirements.txt
tkinter
beautifulsoup4
yt-dlp
requests
pyautogui
pyperclip
google-generativeai
langchain
langchain-community
langchain-core
SQLAlchemy
SpeechRecognition
pyttsx3
pyaudio
pandas
openpyxl
python-docx
python-pptx
PyPDF2
pillow
pymupdf
python-dotenv
pyfiglet
winapps
pywhatkit
smtplib
getpass
mimetypes